fix(transformer): C13 logit_bias 放宽 map[string]float64,修复浮点值解码中断

+++ llm/model.go
@@ -151,7 +151,7 @@ type Request struct {
 	// effect will vary per model, but values between -1 and 1 should decrease or
 	// increase likelihood of selection; values like -100 or 100 should result in a ban
 	// or exclusive selection of the relevant token.
-	LogitBias map[string]int64 `json:"logit_bias,omitempty"`
+	LogitBias map[string]float64 `json:"logit_bias,omitempty"`
 
 	// Set of 16 key-value pairs that can be attached to an object. This can be useful
 	// for storing additional information about the object in a structured format, and
